requirements.txt
requests==2.4.3
responses==0.3.0
pytest==2.6.4
mock==1.0.1
py==1.4.26
python-dateutil==2.3
PyYAML==3.11
